% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_log_parse.R
\name{apache_log_parse}
\alias{apache_log_parse}
\title{apache_log_parse}
\usage{
apache_log_parse(log_file)
}
\arguments{
\item{log_file}{String name/path of log}
}
\value{
log_df Data frame with parsed log columns
}
\description{
Parse Apache-style access log file
}
