% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessor.R
\name{preProcessor}
\alias{preProcessor}
\title{a class to load and preProcess data}
\usage{
preProcessor()
}
\description{
Load and preprocess Data
             \itemize{
             \item \code{preProcessor() } the costructor
             \item \code{load.data.frame( ... ) } Carica il data frame dei referti
             \item \code{getData() } restituisce i dati processati
             \item \code{apply.filter() } applica un filtro ai referti
             }
}
\examples{
\dontrun{
# -----------------------------------------------
#  USING THE METHODS of the class
# -----------------------------------------------
obj.pp <- preProcessor();  

obj.pp$load.data.frame( referti, colonnaTestoReferto="descrizione" )

# Rimuovi i documenti con meno di 10 parole 
obj.pp$apply.filter( filter="remove.little.document", param=list("number.of.words" = 10 )  )

}
}

