% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache_log_parse.R
\name{regex_parse}
\alias{regex_parse}
\title{regex_parse}
\usage{
regex_parse(log_df, log_col)
}
\arguments{
\item{log_df}{Data frame with log}

\item{log_col}{Column name string with log}
}
\value{
log_df_out Data frame with parsed columns
}
\description{
Parse log file string in data frame
}
